Reorganize components in analysis
