+++ src/analysis/dashboard/AnalysisDashboard.tsx
@@ -1,7 +1,7 @@
 import { AppShell, Container } from '@mantine/core';
-import { GlobalConfig } from '../parser/types';
-import { SummaryBlock } from './dashboard/SummaryBlock';
-import AppHeader from './components/interface/AppHeader';
+import { GlobalConfig } from '../../parser/types';
+import { SummaryBlock } from './SummaryBlock';
+import AppHeader from '../interface/AppHeader';
 
 export function AnalysisDashboard({ globalConfig }: { globalConfig: GlobalConfig; }) {
   return (
