+++ src/analysis/stats/AnswerPanel.tsx
@@ -4,7 +4,7 @@ import {
 import React, { useEffect, useMemo, useState } from 'react';
 import { VegaLite, VisualizationSpec } from 'react-vega';
 import { useResizeObserver } from '@mantine/hooks';
-import { IndividualComponent, InheritedComponent, RadioResponse } from '../../../parser/types';
+import { IndividualComponent, InheritedComponent, RadioResponse } from '../../parser/types';
 
 export default function AnswerPanel({ data, config }: { data: Record<string, Record<string, unknown>>, config: IndividualComponent | InheritedComponent | undefined}) {
   const [correctUser, setCorrectUser] = useState<string[]>([]);
@@ -53,7 +53,7 @@ export default function AnswerPanel({ data, config }: { data: Record<string, Rec
       setCorrectUser(correct);
       setIncorrectUser(incorrect);
     }
-  }, [data]);
+  }, [config, data]);
 
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   const specBoxer = useMemo(() => ({
