--- src/analysis/stats/IndividualStats.tsx
@@ -1,5 +1,0 @@
-export default function IndividualStats() {
-  return (
-    <h1>Individual Stats</h1>
-  );
-}
