--- src/analysis/stats/SummaryStats.tsx
@@ -1,5 +1,0 @@
-export default function SummaryStats() {
-  return (
-    <h1>Summary Stats</h1>
-  );
-}
